BackendTJ/requirements.txt
click==7.1.2
dnspython==2.1.0
fastapi==0.65.1
h11==0.12.0
motor==2.3.1
odmantic==0.3.5
pydantic==1.8.2
pymongo==3.11.4
PyYAML==5.4.1
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.13.4
click==7.1.2
dnspython==2.1.0
fastapi==0.65.1
h11==0.12.0
motor==2.3.1
odmantic==0.3.5
pydantic==1.8.2
pymongo==3.11.4
PyYAML==5.4.1
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.13.4
